+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.2
+
+* Modify the 'ChatUIKitTheme.of' return type;
+* Update barrageColor lightness.
+
 ## 0.0.1
 
 * Add light and dark theme;